TST/REF: collect tests by method from tests.generic
